convert Picasso output p2e_go to earlinet data format using cloud_info.nc
the three scripts use the profiles.nc and the cloud_info.nc to convert Picasso output with p2e_go to earlinet data format. As upper height for the profile, the cloud base - 500 m is used. Further imporvements are needed. Like the implementation of OC files. Thus it would be beneficial to impelment p2e not as executable but a spythn script to the bash script. Also path names need top be made variable.